Performed cleanup
