+++ packages/client/src/tests/mock-offline-data.ts
@@ -15,7 +15,7 @@ import { join } from 'path'
 import {
   IFormDraft,
   DraftStatus
-} from '@client/forms/configuration/formDrafts/reducer'
+} from '@client/forms/configuration/formDrafts/utils'
 
 export const validImageB64String =
   'iVBORw0KGgoAAAANSUhEUgAAAAgAAAACCAYAAABllJ3tAAAABHNCSVQICAgIfAhkiAAAABl0RVh0U29mdHdhcmUAZ25vbWUtc2NyZWVuc2hvdO8Dvz4AAAAXSURBVAiZY1RWVv7PgAcw4ZNkYGBgAABYyAFsic1CfAAAAABJRU5ErkJggg=='
